Extended scripts in ScriptedModel by an optional function setupEnvironment(), in which collision rules and later also global values and randomization can be defined. This method is called after createEnvironment, when all new objects are registered with their valid names at the PhysicsManager, so that they can be adressed via name.
Modified all DebugPanels (ValueDetailPanel, EventDetailPanel, EventVisualizationWindow, ...) so that they can be used like the propety panel: ** is the same as .*, the entries are not case sensitive any more.
ValueDetailPanel: Added the new description property of values to the panel, so that the panel can be used as part of the documentation.
Extended ValueManager and EventManager by getEvents/getValues methods that allow the specification of whether the regexp. should be case sensitive or insensitive.