Modified local evolution settings override: the local deviation, min, max, and change probability can now be given as - value (e.g. 0.01) - percentage (e.g. %10, 250%) - variable (e.g. $propertyNameOfNetwork)
Note that variables refer to properties of the network, so you have to add a valid property to the network.

Renamed SYNAPSE_DEVIATION and SYNAPSE_CHANGE_PROBABILITY to ELEMENT_DEVIATION and ELEMENT_CHANGE_PROB (also the editor names of the tags have been changed.)
Added NeuroEvolutionUtil with a central place to to variable interpretation.