Refactoring to synchronize the access of the matrix values of the plotters. DynamicPlotManager: now provides a mutex that has to be locked whenever a matrix (mData of the plotters9 is read or written. MatlabExporter: locks the matrixMutex when accessing a matrix to export. OnlinePlotterWindow: locks the matrixMutex when copying the matrix for local plotting. Matrix: A copied matrix automatically forces a deep copy of the QVectors (that usually do implicit sharing). This sharing is not thread save, so the copy is enforced (by using non-const functions of all vectors.) All Plotters: locking of the matrixMutex when changing and accessing the data matrix.
main: Is now also prepared for Windows and initializes the XWindows system.
CMAKE: removed superfluous definitions of the build type, so that only one file has to be changed to switch between release and debug.

TODO: check the impact of this modification on performance. Maybe locking should be optimized to minimize the number of locks to gain a higher performance.